Move JSON parameter parsing into separate class
Fixes #8904

This is so that it can be reused and/or the implementation can be
changed more essily.
